requirements.txt
streamlit
streamlit-auth
pyyaml
unittest2
pandas